Corrections on many files.
changelog:
- corrections on many files.
