+++ package.json
@@ -40,8 +40,6 @@
   "dependencies": {
     "axios": "0.27.2",
     "form-data": "4.0.0",
-    "jsdom": "^19.0.0",
-    "open": "^8.4.0",
     "mime": "^3.0.0",
     "skynet-js": "^4.3.0",
     "@skynetlabs/tus-js-client": "^2.4.1",
