+++ scripts/download.js
@@ -24,7 +24,7 @@ if (!fs.existsSync(dir)) {
   let usedSkylink;
   let usedHnsLink;
 
-  if ((process.argv[2] === null) | (process.argv[2] === undefined)) {
+  if (process.argv[2] === null || process.argv[2] === undefined) {
     usedSkylink = defaultSkylink;
     console.log("\n\n\nusedSkylink =  " + usedSkylink);
     usedHnsLink = defaultHnsLink;
@@ -40,9 +40,8 @@ if (!fs.existsSync(dir)) {
   async function downloadFile(path, skylink) {
     await client
       .downloadFile(path, skylink)
-      .then((res) => {
+      .then(() => {
         console.log("\n\n\n1. use downloadFile to get a file from skynet.");
-        console.log("skylinkUrl: " + res);
       })
       .catch((err) => {
         console.log("\n1. Get Error: ", err);
@@ -54,7 +53,7 @@ if (!fs.existsSync(dir)) {
     await client
       .downloadFileHns(path, hnsLink)
       .then((res) => {
-        console.log("\n\n\n2. use downloadFileHns to get a hns file from skynet.");
+        console.log("\n\n2. use downloadFileHns to get a hns file from skynet.");
         console.log("hnsLink: " + res);
       })
       .catch((err) => {
@@ -67,7 +66,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getSkylinkUrl(skylink)
       .then((res) => {
-        console.log("\n\n\n3. use getSkylinkUrl to get the url from a skylink.");
+        console.log("\n\n3. use getSkylinkUrl to get the url from a skylink.");
         console.log("skylinkUrl: " + res);
       })
       .catch((err) => {
@@ -80,7 +79,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getHnsUrl(hnsLink)
       .then((res) => {
-        console.log("\n\n\n4. use getHnsUrl to get the hnsUrl from a hnsLink.");
+        console.log("\n\n4. use getHnsUrl to get the hnsUrl from a hnsLink.");
         console.log("hnsUrl: " + res);
       })
       .catch((err) => {
@@ -93,7 +92,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getHnsresUrl(hnsLink)
       .then((res) => {
-        console.log("\n\n\n5. use getHnsresUrl to get the hnsresUrl from a hnsLink.");
+        console.log("\n\n5. use getHnsresUrl to get the hnsresUrl from a hnsLink.");
         console.log("hnsresUrl: " + res);
       })
       .catch((err) => {
@@ -106,7 +105,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getMetadata(skylink)
       .then((res) => {
-        console.log("\n\n\n6. use getMetadata to get the metadata from a skylink.");
+        console.log("\n\n6. use getMetadata to get the metadata from a skylink.");
         console.log("Metadata: " + JSON.stringify(res));
       })
       .catch((err) => {
@@ -119,8 +118,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getFileContent(entryLink)
       .then((res) => {
-        console.log("\n\n\n7. use getFileContent to get the fileContent from a skylink.");
-        console.log("data: " + res.data);
+        console.log("\n\n7. use getFileContent to get the fileContent from a skylink.");
         console.log("contentType: " + res.contentType);
         console.log("portalUrl: " + res.portalUrl);
         console.log("skylink: " + res.skylink);
@@ -135,8 +133,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getFileContentBinary(skylink)
       .then((res) => {
-        console.log("\n\n\n8. use getFileContentBinary to get the fileContentBinary from a skylink.");
-        console.log("binaryData: " + res.data);
+        console.log("\n\n8. use getFileContentBinary to get the fileContentBinary from a skylink.");
         console.log("contentType: " + res.contentType);
         console.log("portalUrl: " + res.portalUrl);
         console.log("skylink: " + res.skylink);
@@ -151,8 +148,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getFileContentHns(hnsLink)
       .then((res) => {
-        console.log("\n\n\n9. use getFileContentHns to get the hns fileContent from a hnsLink.");
-        console.log("data: " + res.data);
+        console.log("\n\n9. use getFileContentHns to get the hns fileContent from a hnsLink.");
         console.log("contentType: " + res.contentType);
         console.log("portalUrl: " + res.portalUrl);
         console.log("skylink: " + res.skylink);
@@ -167,8 +163,7 @@ if (!fs.existsSync(dir)) {
     await client
       .getFileContentBinaryHns(hnsLink)
       .then((res) => {
-        console.log("\n\n\n10. use getFileContentBinaryHns to get the hnsFileContentBinary from a hnsLink.");
-        console.log("binaryData: " + res.data);
+        console.log("\n\n10. use getFileContentBinaryHns to get the hnsFileContentBinary from a hnsLink.");
         console.log("contentType: " + res.contentType);
         console.log("portalUrl: " + res.portalUrl);
         console.log("skylink: " + res.skylink);
@@ -178,43 +173,17 @@ if (!fs.existsSync(dir)) {
       });
   }
 
-  //11. use openFile to open a skylink in browser.
-  async function openFile(skylink) {
-    await client
-      .openFile(skylink)
-      .then((res) => {
-        console.log("\n\n\n11. use openFile to open a skylink in browser.");
-        console.log("skylinkUrl: " + res);
-      })
-      .catch((err) => {
-        console.log("\n11. Get Error: ", err);
-      });
-  }
-
-  //12. use openFileHns to open a hnsLink in browser.
-  async function openFileHns(hnsLink) {
-    await client
-      .openFileHns(hnsLink)
-      .then((res) => {
-        console.log("\n\n\n12. use openFileHns to open a hnsLink in browser.");
-        console.log("hnsUrl: " + res);
-      })
-      .catch((err) => {
-        console.log("\n12. Get Error: ", err);
-      });
-  }
-
-  //13. use resolveHns to get from a hnsLink the data.
+  //11. use resolveHns to get from a hnsLink the data.
   async function resolveHns(hnsLink) {
     await client
       .resolveHns(hnsLink)
       .then((res) => {
-        console.log("\n\n\n13. use resolveHns to get from a hnsLink the data.");
+        console.log("\n\n11. use resolveHns to get from a hnsLink the data.");
         console.log("data: " + JSON.stringify(res.data));
         console.log("skylink: " + res.skylink);
       })
       .catch((err) => {
-        console.log("\n13. Get Error: ", err);
+        console.log("\n11. Get Error: ", err);
       });
   }
 
@@ -229,8 +198,6 @@ if (!fs.existsSync(dir)) {
     await getFileContentBinary(usedSkylink);
     await getFileContentHns(usedHnsLink);
     await getFileContentBinaryHns(usedHnsLink);
-    await openFile(usedSkylink);
-    await openFileHns(usedHnsLink);
     await resolveHns(usedHnsLink);
   }
   main();
